epic_games_bot/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
python-telegram-bot>=13.0.0
discord.py>=2.0.0
